Fix queue build up issue in gpsd_client
In step() which is called periodically according to the parameter
publish_rate we only perform read() once, which means that there could
still be more unprocessed data from gpsd waiting. This leads to a build
up of unprocessed data causing the published data to be old and
outdated.

By instead using the waiting() call to see if there's more
data to be processed we can make sure that what we publish is actually
the newest available data.